fix : inactive 티켓 선택하는 모달을 app2에서 보여주는 것으로 수정
